backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic>=2.8.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
sqlalchemy>=2.0.31
alembic==1.12.1
# psycopg2-binary==2.9.9  # PostgreSQL için, SQLite kullanıyorsak gerekli değil
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow>=11.0.0
routeros-api==0.19.0
aiofiles==23.2.1
aiosqlite==0.19.0
slowapi==0.1.9
user-agents==2.2.0

